Split API blueprints into separate files
